Split Chart + SeriesAreaManager code.
